static_cast<int> removed

+++ Workshop 2/barbarian.h
@@ -29,7 +29,7 @@ namespace seneca {
 
 	template <typename T, typename Ability_t, typename Weapon_t>
 	int Barbarian<T, Ability_t, Weapon_t>::getAttackAmnt() const {
-		return (m_baseAttack + (static_cast<int>(m_weapon[0]) / 2) + (static_cast<int>(m_weapon[1]) / 2));
+		return static_cast<int>(m_baseAttack + (m_weapon[0] / 2) + m_weapon[1]) / 2);
 	}
 
 	template <typename T, typename Ability_t, typename Weapon_t>
